feat: login and user validation on actions

+++ src/graphql/mutations/task/task-destroy-mutation.ts
@@ -1,4 +1,8 @@
+import { ValidationError } from 'apollo-server-errors';
 import { mutationField, nonNull } from 'nexus';
+
+import { getUserId } from 'graphql/utils/auth-helpers';
+import lang from 'lang';
 
 export const TaskDestroyMutation = mutationField('taskDestroy', {
   type: nonNull('TaskSingleResult'),
@@ -6,6 +10,8 @@ export const TaskDestroyMutation = mutationField('taskDestroy', {
     taskId: nonNull('UUID'),
   },
   resolve: async (parent, { taskId }, ctx) => {
+    const user = await getUserId(ctx);
+
     const task = await ctx.prisma.task.findUnique({
       where: {
         id: taskId,
@@ -13,7 +19,11 @@ export const TaskDestroyMutation = mutationField('taskDestroy', {
     });
 
     if (!task) {
-      throw new Error(`Task not found`);
+      throw new ValidationError(lang.task.not_found);
+    }
+
+    if (task.userId !== user.id) {
+      throw new ValidationError(lang.permission.action_not_allowed);
     }
 
     const data = await ctx.prisma.task.delete({
